pyproject.toml
[project]
name = "stopdb"
version = "0.1.0"
description = "Backend database and administrative views for Stop Hate In School."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1",
    "psycopg>=3.2.1",
    "dj-database-url>=2.2.0",
    "whitenoise>=6.7.0",
    "django-phonenumber-field[phonenumberslite]>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.2",
    "mypy>=1.11.1",
    "django-stubs[compatible-mypy]>=5.0.4",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "server.settings"
